fix mysql copy

+++ docker/mysql/export_common.sh
@@ -8,6 +8,7 @@ export WALG_STREAM_CREATE_COMMAND="xtrabackup --backup \
     --host=localhost \
     --parallel=2 \
     --datadir=${MYSQLDATA}"
+
 export WALG_STREAM_RESTORE_COMMAND="xbstream -x -C ${MYSQLDATA}"
 export WALG_MYSQL_BACKUP_PREPARE_COMMAND="xtrabackup --prepare --target-dir=${MYSQLDATA}"
 export WALG_MYSQL_BINLOG_REPLAY_COMMAND='mysqlbinlog --stop-datetime="$WALG_MYSQL_BINLOG_END_TS" "$WALG_MYSQL_CURRENT_BINLOG" | mysql'
@@ -15,8 +16,8 @@ export WALG_MYSQL_BINLOG_DST=/tmp
 
 # test tools
 mysql_kill_and_clean_data() {
-    kill -9 `pidof mysqld` || true
-    rm -rf "${MYSQLDATA}"/*
+    kill -9 "$(pidof mysqld)" || true
+    rm -rf "${MYSQLDATA:?}"
     rm -rf /root/.walg_mysql_binlogs_cache
 }
 
@@ -27,7 +28,7 @@ mysql_set_gtid_purged() {
 }
 
 sysbench() {
-    /usr/bin/sysbench --verbosity=0 /usr/share/sysbench/oltp_insert.lua $@
+    /usr/bin/sysbench --verbosity=0 /usr/share/sysbench/oltp_insert.lua "$@"
 }
 
 date3339() {
